+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file contains the list of changes made to pyjoulescope.
 
 *   Added joulescope.inf to manually force Win 7 machines without WCID update 
     to recognize Joulescopes.
+*   Fixed divide by zero error in stream_buffer.stats_compute_end.
 
 
 ## 0.4.4